Delete task warning

+++ public/main.js
@@ -14,13 +14,16 @@ Array.from(edit).forEach(function (element) {
 Array.from(trash).forEach(function (element) {
   element.addEventListener('click', function () {
     const _id = this.closest('li').querySelector('span').getAttribute("name")
+    // confirmation before deleting
+    const confirmed = window.confirm("Are you sure you want to delete this task? This action cannot be undone.");
+    if (!confirmed) return;
+
     fetch(`/task/${_id}`, {
       method: 'delete',
       headers: {
         'Content-Type': 'application/json'
       },
-      body: JSON.stringify({
-      })
+      body: JSON.stringify({})
     }).then(function (response) {
       window.location.reload()
     })
